Start making deferrable analytics possible in client
