+++ internal/captain/command.go
@@ -400,15 +400,15 @@ func (c *Command) subCommandNames() []string {
 }
 
 func (c *Command) runner(cobraCmd *cobra.Command, args []string) error {
+	analytics.SetDefer(c.deferAnalytics)
+
 	outputFlag := cobraCmd.Flag("output")
 	if outputFlag != nil && outputFlag.Changed {
 		analytics.SetDimensionsOutput(outputFlag.Value.String())
 	}
 	subCommandString := c.UseFull()
-	// Send  GA events unless they are handled in the runners...
-	if !c.deferAnalytics {
-		analytics.Event(analytics.CatRunCmd, subCommandString)
-	}
+	analytics.Event(analytics.CatRunCmd, subCommandString)
+
 	// Run OnUse functions for non-persistent flags
 	c.runFlags(false)
 
@@ -444,10 +444,10 @@ func (c *Command) runner(cobraCmd *cobra.Command, args []string) error {
 	execute := intercept(c.execute)
 
 	err := execute(c, args)
-	if !c.deferAnalytics {
-		exitCode := errs.UnwrapExitCode(failures.ToError(err))
-		analytics.EventWithLabel(analytics.CatCommandExit, subCommandString, strconv.Itoa(exitCode))
-	}
+
+	exitCode := errs.UnwrapExitCode(failures.ToError(err))
+	analytics.EventWithLabel(analytics.CatCommandExit, subCommandString, strconv.Itoa(exitCode))
+
 	return err
 }
